requirements.txt
opencv-python
ffmpeg-python
torch
numpy
tqdm
moviepy
openai-whisper
openai
python-dotenv
mediapipe
scikit-image